refactor(project): split settings source resolution
